chore: move utils away from action file
