+++ packages/client/src/v2-events/features/events/useEvents/procedures/actions/action.ts
@@ -9,7 +9,6 @@
  * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
  */
 
-import { Mutation as TanstackMutation } from '@tanstack/query-core'
 import { useMutation } from '@tanstack/react-query'
 import {
   DecorateMutationProcedure,
@@ -17,108 +16,21 @@ import {
   inferOutput
 } from '@trpc/tanstack-react-query'
 import {
-  ActionInput,
   ActionType,
-  EventDocument,
-  getCurrentEventState,
   stripHiddenOrDisabledFields
 } from '@opencrvs/commons/client'
 import { useEventConfigurations } from '@client/v2-events/features/events/useEventConfiguration'
 import {
-  invalidateEventsList,
-  setEventData,
-  setEventListData,
-  findLocalEventData
+  findLocalEventData,
+  updateLocalEvent
 } from '@client/v2-events/features/events/useEvents/api'
 import { queryClient, trpcOptionsProxy } from '@client/v2-events/trpc'
-import { createTemporaryId } from '@client/v2-events/utils'
 import * as customApi from '@client/v2-events/custom-api'
-import { setMutationDefaults, waitUntilEventIsCreated } from './utils'
-
-async function updateLocalEvent(updatedEvent: EventDocument) {
-  setEventData(updatedEvent.id, updatedEvent)
-  return invalidateEventsList()
-}
-
-/*
- * This makes sure that if you are offline and do
- * 1. Create record
- * 2. Create draft
- * 3. Declare the record
- * 4. Connect to the internet
- * The draft stage in the middle will be cancelled. This is to prevent race conditions
- * between when the backend receives the draft and when it receives the declare action.
- */
-function cancelOngoingDraftRequests({ eventId, draft }: ActionInput) {
-  const mutationCache = queryClient.getMutationCache()
-
-  const isDraftMutation = (
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    mutation: TanstackMutation<unknown, Error, any, unknown>
-  ): mutation is TanstackMutation<unknown, Error, ActionInput, unknown> => {
-    const mutationKey = mutation.options.mutationKey
-    if (!mutationKey || !mutationKey[0]) {
-      return false
-    }
-    return ['event', 'actions'].every((key) => mutationKey.flat().includes(key))
-  }
-
-  const draftMutationsForThisEvent = mutationCache
-    .getAll()
-    .filter(
-      (mutation) =>
-        isDraftMutation(mutation) &&
-        mutation.state.variables?.eventId === eventId &&
-        mutation.state.variables.draft
-    )
-
-  if (!draft) {
-    draftMutationsForThisEvent.forEach((mutation) => {
-      mutationCache.remove(mutation)
-    })
-  } else {
-    // Keep the last draft mutation as it's this current request
-    draftMutationsForThisEvent.slice(0, -1).forEach((mutation) => {
-      mutationCache.remove(mutation)
-    })
-  }
-}
-
-function updateEventOptimistically<T extends ActionInput>(
-  actionType: typeof ActionType.DECLARE
-) {
-  return (variables: T) => {
-    cancelOngoingDraftRequests(variables)
-
-    const localEvent = queryClient.getQueryData(
-      trpcOptionsProxy.event.get.queryKey(variables.eventId)
-    )
-    if (!localEvent) {
-      return
-    }
-    const optimisticEvent: EventDocument = {
-      ...localEvent,
-      actions: [
-        ...localEvent.actions,
-        {
-          id: createTemporaryId(),
-          type: actionType,
-          data: variables.data,
-          draft: false,
-          createdAt: new Date().toISOString(),
-          createdBy: '@todo',
-          createdAtLocation: '@todo'
-        }
-      ]
-    }
-
-    setEventListData((eventIndices) =>
-      eventIndices
-        ?.filter((ei) => ei.id !== optimisticEvent.id)
-        .concat(getCurrentEventState(optimisticEvent))
-    )
-  }
-}
+import { updateEventOptimistically } from '@client/v2-events/features/events/useEvents/procedures/actions/utils'
+import {
+  waitUntilEventIsCreated,
+  setMutationDefaults
+} from '@client/v2-events/features/events/useEvents/procedures/utils'
 
 setMutationDefaults(trpcOptionsProxy.event.actions.declare, {
   mutationFn: createEventActionMutationFn(
