+++ packages/client/src/v2-events/features/events/useEvents/procedures/actions/utils.ts
@@ -1,0 +1,101 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * OpenCRVS is also distributed under the terms of the Civil Registration
+ * & Healthcare Disclaimer located at http://opencrvs.org/license.
+ *
+ * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
+ */
+
+import { Mutation as TanstackMutation } from '@tanstack/query-core'
+import {
+  ActionInput,
+  ActionType,
+  EventDocument,
+  getCurrentEventState
+} from '@opencrvs/commons/client'
+import { setEventListData } from '@client/v2-events/features/events/useEvents/api'
+import { queryClient, trpcOptionsProxy } from '@client/v2-events/trpc'
+import { createTemporaryId } from '@client/v2-events/utils'
+
+/*
+ * This makes sure that if you are offline and do
+ * 1. Create record
+ * 2. Create draft
+ * 3. Declare the record
+ * 4. Connect to the internet
+ * The draft stage in the middle will be cancelled. This is to prevent race conditions
+ * between when the backend receives the draft and when it receives the declare action.
+ */
+export function cancelOngoingDraftRequests({ eventId, draft }: ActionInput) {
+  const mutationCache = queryClient.getMutationCache()
+
+  const isDraftMutation = (
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    mutation: TanstackMutation<unknown, Error, any, unknown>
+  ): mutation is TanstackMutation<unknown, Error, ActionInput, unknown> => {
+    const mutationKey = mutation.options.mutationKey
+    if (!mutationKey || !mutationKey[0]) {
+      return false
+    }
+    return ['event', 'actions'].every((key) => mutationKey.flat().includes(key))
+  }
+
+  const draftMutationsForThisEvent = mutationCache
+    .getAll()
+    .filter(
+      (mutation) =>
+        isDraftMutation(mutation) &&
+        mutation.state.variables?.eventId === eventId &&
+        mutation.state.variables.draft
+    )
+
+  if (!draft) {
+    draftMutationsForThisEvent.forEach((mutation) => {
+      mutationCache.remove(mutation)
+    })
+  } else {
+    // Keep the last draft mutation as it's this current request
+    draftMutationsForThisEvent.slice(0, -1).forEach((mutation) => {
+      mutationCache.remove(mutation)
+    })
+  }
+}
+
+export function updateEventOptimistically<T extends ActionInput>(
+  actionType: typeof ActionType.DECLARE
+) {
+  return (variables: T) => {
+    cancelOngoingDraftRequests(variables)
+
+    const localEvent = queryClient.getQueryData(
+      trpcOptionsProxy.event.get.queryKey(variables.eventId)
+    )
+    if (!localEvent) {
+      return
+    }
+    const optimisticEvent: EventDocument = {
+      ...localEvent,
+      actions: [
+        ...localEvent.actions,
+        {
+          id: createTemporaryId(),
+          type: actionType,
+          data: variables.data,
+          draft: false,
+          createdAt: new Date().toISOString(),
+          createdBy: '@todo',
+          createdAtLocation: '@todo'
+        }
+      ]
+    }
+
+    setEventListData((eventIndices) =>
+      eventIndices
+        ?.filter((ei) => ei.id !== optimisticEvent.id)
+        .concat(getCurrentEventState(optimisticEvent))
+    )
+  }
+}
