+++ packages/client/src/v2-events/features/events/useEvents/useEvents.ts
@@ -16,7 +16,7 @@ import { useOutbox } from './outbox'
 import { useCreateEvent } from './procedures/create'
 import { useDeleteEvent } from './procedures/delete'
 import { useGetEvent } from './procedures/get'
-import { customMutationKeys, useEventAction } from './procedures/action'
+import { customMutationKeys, useEventAction } from './procedures/actions/action'
 
 export function useEvents() {
   const trpc = useTRPC()
